refactor: consolidate command interface surface
