+++ src/__tests__/cli-grammar.test.ts
@@ -1,13 +1,7 @@
 import { test } from 'vitest';
 import assert from 'node:assert/strict';
 import { DAEMON_COMMAND_GROUPS, PUBLIC_COMMANDS } from '../command-catalog.ts';
-import {
-  readFillTargetFromPositionals,
-  readInteractionTargetFromPositionals,
-  readLongPressTargetFromPositionals,
-  readSemanticInputFromCli,
-  readWaitOptionsFromPositionals,
-} from '../commands/semantic-grammar.ts';
+import { readInputFromCli } from '../commands/cli-grammar.ts';
 import type { CliFlags } from '../utils/command-schema.ts';
 
 const BASE_FLAGS: CliFlags = {
@@ -23,44 +17,41 @@ test('command catalog owns daemon routing groups', () => {
 });
 
 test('wait grammar preserves CLI bare text forms', () => {
-  const options = readWaitOptionsFromPositionals(['Continue', '1500'], BASE_FLAGS);
+  const options = readInputFromCli('wait', ['Continue', '1500'], BASE_FLAGS);
   assert.equal(options.text, 'Continue');
   assert.equal(options.timeoutMs, 1500);
 });
 
 test('interaction and fill grammar share ref, selector, and point parsing', () => {
-  assert.deepEqual(readInteractionTargetFromPositionals(['@e3', 'Email']), {
+  assert.deepEqual(readInputFromCli('press', ['@e3', 'Email'], BASE_FLAGS).target, {
+    kind: 'ref',
     ref: '@e3',
     label: 'Email',
   });
-  assert.deepEqual(readFillTargetFromPositionals(['id=email', 'qa@example.com']), {
-    kind: 'selector',
-    target: { selector: 'id=email' },
-    text: 'qa@example.com',
-  });
-  assert.deepEqual(readFillTargetFromPositionals(['@e4', 'Email', 'qa@example.com']), {
-    kind: 'ref',
-    target: { ref: '@e4', label: 'Email' },
-    text: 'qa@example.com',
-  });
-  assert.deepEqual(readFillTargetFromPositionals(['10', '20', 'hello']), {
-    kind: 'point',
-    target: { x: 10, y: 20 },
-    text: 'hello',
-  });
-  assert.deepEqual(readLongPressTargetFromPositionals(['@e4', '800']), {
-    ref: '@e4',
+  const selectorFill = readInputFromCli('fill', ['id=email', 'qa@example.com'], BASE_FLAGS);
+  assert.deepEqual(selectorFill.target, { kind: 'selector', selector: 'id=email' });
+  assert.equal(selectorFill.text, 'qa@example.com');
+
+  const refFill = readInputFromCli('fill', ['@e4', 'Email', 'qa@example.com'], BASE_FLAGS);
+  assert.deepEqual(refFill.target, { kind: 'ref', ref: '@e4', label: 'Email' });
+  assert.equal(refFill.text, 'qa@example.com');
+
+  const pointFill = readInputFromCli('fill', ['10', '20', 'hello'], BASE_FLAGS);
+  assert.deepEqual(pointFill.target, { kind: 'point', x: 10, y: 20 });
+  assert.equal(pointFill.text, 'hello');
+
+  assert.deepEqual(readInputFromCli('longpress', ['@e4', '800'], BASE_FLAGS), {
+    target: { kind: 'ref', ref: '@e4' },
     durationMs: 800,
   });
-  assert.deepEqual(readLongPressTargetFromPositionals(['10', '20', '800']), {
-    x: 10,
-    y: 20,
+  assert.deepEqual(readInputFromCli('longpress', ['10', '20', '800'], BASE_FLAGS), {
+    target: { kind: 'point', x: 10, y: 20 },
     durationMs: 800,
   });
 });
 
 test('find and is grammar decodes command action positionals', () => {
-  const findOptions = readSemanticInputFromCli('find', ['label', 'Continue', 'wait', '3000'], {
+  const findOptions = readInputFromCli('find', ['label', 'Continue', 'wait', '3000'], {
     ...BASE_FLAGS,
     platform: 'ios',
     findFirst: true,
@@ -72,21 +63,17 @@ test('find and is grammar decodes command action positionals', () => {
   assert.equal(findOptions.timeoutMs, 3000);
   assert.equal(findOptions.first, true);
 
-  const isOptions = readSemanticInputFromCli('is', ['text', 'id=title', 'Welcome'], BASE_FLAGS);
+  const isOptions = readInputFromCli('is', ['text', 'id=title', 'Welcome'], BASE_FLAGS);
   assert.equal(isOptions.predicate, 'text');
   assert.equal(isOptions.selector, 'id=title');
   assert.equal(isOptions.value, 'Welcome');
 });
 
 test('settings grammar owns positional parsing for CLI commands', () => {
-  const location = readSemanticInputFromCli(
-    'settings',
-    ['location', 'set', '37.3349', '-122.009'],
-    {
-      ...BASE_FLAGS,
-      platform: 'ios',
-    },
-  );
+  const location = readInputFromCli('settings', ['location', 'set', '37.3349', '-122.009'], {
+    ...BASE_FLAGS,
+    platform: 'ios',
+  });
   assert.equal(location.platform, 'ios');
   assert.equal(location.setting, 'location');
   assert.equal(location.state, 'set');
