+++ src/commands/cli-grammar.ts
@@ -1,0 +1,8 @@
+export {
+  prepareBatchStep,
+  prepareDaemonCommandRequest,
+  readInputFromCli,
+  batchCommandNames,
+  type BatchCommandName,
+  type DaemonCommandName,
+} from './cli-grammar/registry.ts';
