+++ src/commands/interaction-command-contracts.ts
@@ -1,17 +1,27 @@
 import type {
+  AgentDeviceClient,
   ClickOptions,
+  FindOptions,
   FlingOptions,
   FillOptions,
+  FocusOptions,
+  GetOptions,
   PanOptions,
   PinchOptions,
   PressOptions,
+  IsOptions,
+  LongPressOptions,
   RotateGestureOptions,
+  ScrollOptions,
+  SwipeOptions,
   TransformGestureOptions,
+  TypeTextOptions,
 } from '../client-types.ts';
-import { defineSemanticCommand } from './semantic-contract.ts';
-import { tapCliOutput } from './semantic-client-output.ts';
+import { defineCommand } from './command-contract.ts';
 import {
+  booleanField,
   commonToClientOptions,
+  elementTargetField,
   enumField,
   fieldsInputSchema,
   integerField,
@@ -29,17 +39,32 @@ import {
   requiredNumber,
   selectorSnapshotFields,
   stringField,
+  toClientElementTarget,
   toClientInteractionTarget,
   toRepeatedOptions,
   toSelectorSnapshotOptions,
   type CommonCommandInput,
+  type CommandFieldMap,
   type InferCommandInput,
   type PointInput,
-} from './semantic-common.ts';
+} from './command-input.ts';
 
 const CLICK_BUTTON_VALUES = ['primary', 'secondary', 'middle'] as const;
 const GESTURE_KIND_VALUES = ['pan', 'fling', 'pinch', 'rotate', 'transform'] as const;
 const GESTURE_DIRECTION_VALUES = ['up', 'down', 'left', 'right'] as const;
+const FIND_ACTION_VALUES = [
+  'click',
+  'focus',
+  'exists',
+  'getText',
+  'getAttrs',
+  'wait',
+  'fill',
+  'type',
+] as const;
+const FIND_LOCATOR_VALUES = ['any', 'text', 'label', 'value', 'role', 'id'] as const;
+const SCROLL_DIRECTION_VALUES = ['up', 'down', 'left', 'right', 'top', 'bottom'] as const;
+const SWIPE_PATTERN_VALUES = ['one-way', 'ping-pong'] as const;
 
 const clickFields = {
   target: requiredField(interactionTargetField()),
@@ -64,6 +89,64 @@ const fillFields = {
   ...selectorSnapshotFields(),
 };
 
+const longPressFields = {
+  target: requiredField(interactionTargetField()),
+  durationMs: integerField('Long press duration in milliseconds.', { min: 0 }),
+  ...selectorSnapshotFields(),
+};
+
+const swipeFields = {
+  from: requiredField(pointField('Swipe start point.')),
+  to: requiredField(pointField('Swipe end point.')),
+  durationMs: integerField('Swipe duration in milliseconds.', { min: 0 }),
+  count: integerField('Number of swipe repetitions.', { min: 1 }),
+  pauseMs: integerField('Pause between repeated swipes.', { min: 0 }),
+  pattern: enumField(SWIPE_PATTERN_VALUES),
+};
+
+const focusFields = {
+  x: requiredField(numberField('X coordinate.')),
+  y: requiredField(numberField('Y coordinate.')),
+};
+
+const typeFields = {
+  text: requiredField(stringField('Text to type.')),
+  delayMs: integerField('Delay between typed characters.', { min: 0 }),
+};
+
+const scrollFields = {
+  direction: requiredField(enumField(SCROLL_DIRECTION_VALUES)),
+  amount: numberField('Platform scroll amount.'),
+  pixels: integerField('Pixel scroll amount.', { min: 0 }),
+};
+
+const getFields = {
+  format: requiredField(enumField(['text', 'attrs'] as const)),
+  target: requiredField(elementTargetField()),
+  ...selectorSnapshotFields(),
+};
+
+const isFields = {
+  predicate: requiredField(
+    enumField(['visible', 'hidden', 'exists', 'editable', 'selected', 'text'] as const),
+  ),
+  selector: requiredField(stringField()),
+  value: stringField(),
+  ...selectorSnapshotFields(),
+};
+
+const findFields = {
+  locator: enumField(FIND_LOCATOR_VALUES),
+  query: requiredField(stringField()),
+  action: enumField(FIND_ACTION_VALUES),
+  value: stringField(),
+  timeoutMs: integerField(),
+  first: booleanField(),
+  last: booleanField(),
+  depth: integerField(),
+  raw: booleanField(),
+};
+
 const gestureFields = {
   kind: requiredField(enumField(GESTURE_KIND_VALUES, 'Gesture variant.')),
   direction: enumField(GESTURE_DIRECTION_VALUES, 'Fling direction.'),
@@ -79,6 +162,8 @@ const gestureFields = {
 type ClickInput = InferCommandInput<typeof clickFields>;
 type PressInput = InferCommandInput<typeof pressFields>;
 type FillInput = InferCommandInput<typeof fillFields>;
+type LongPressInput = InferCommandInput<typeof longPressFields>;
+type GetInput = InferCommandInput<typeof getFields>;
 
 type PanInput = CommonCommandInput & {
   kind: 'pan';
@@ -119,31 +204,62 @@ type TransformInput = CommonCommandInput & {
 
 type GestureInput = PanInput | FlingInput | PinchInput | RotateInput | TransformInput;
 
-export const interactionSemanticCommands = [
-  defineSemanticCommand({
+export const interactionCommandDefinitions = [
+  defineCommand({
     name: 'click',
     description: 'Click or tap a semantic UI target by ref, selector, or point.',
     inputSchema: fieldsInputSchema(clickFields),
     readInput: (input) => readFieldInput(input, clickFields),
     run: (client, input) => client.interactions.click(toClickOptions(input)),
-    formatCliOutput: ({ result }) => tapCliOutput(result),
   }),
-  defineSemanticCommand({
+  defineCommand({
     name: 'press',
     description: 'Press a semantic UI target by ref, selector, or point.',
     inputSchema: fieldsInputSchema(pressFields),
     readInput: (input) => readFieldInput(input, pressFields),
     run: (client, input) => client.interactions.press(toPressOptions(input)),
-    formatCliOutput: ({ result }) => tapCliOutput(result),
   }),
-  defineSemanticCommand({
+  defineCommand({
     name: 'fill',
     description: 'Fill text into a semantic UI target by ref, selector, or point.',
     inputSchema: fieldsInputSchema(fillFields),
     readInput: (input) => readFieldInput(input, fillFields),
     run: (client, input) => client.interactions.fill(toFillOptions(input)),
   }),
-  defineSemanticCommand({
+  defineFieldCommand(
+    'longpress',
+    'Long press by ref, selector, or point.',
+    longPressFields,
+    (client, input) => client.interactions.longPress(toLongPressOptions(input)),
+  ),
+  defineFieldCommand('swipe', 'Swipe between two points.', swipeFields, (client, input) =>
+    client.interactions.swipe(input as SwipeOptions),
+  ),
+  defineFieldCommand('focus', 'Focus input at coordinates.', focusFields, (client, input) =>
+    client.interactions.focus(input as FocusOptions),
+  ),
+  defineFieldCommand('type', 'Type text in the focused field.', typeFields, (client, input) =>
+    client.interactions.type(input as TypeTextOptions),
+  ),
+  defineFieldCommand(
+    'scroll',
+    'Scroll in a direction or to an edge.',
+    scrollFields,
+    (client, input) => client.interactions.scroll(input as ScrollOptions),
+  ),
+  defineFieldCommand('get', 'Get element text or attributes.', getFields, (client, input) =>
+    client.interactions.get(toGetOptions(input)),
+  ),
+  defineFieldCommand('is', 'Assert UI state.', isFields, (client, input) =>
+    client.interactions.is(input as IsOptions),
+  ),
+  defineFieldCommand(
+    'find',
+    'Find an element and optionally act on it.',
+    findFields,
+    (client, input) => client.interactions.find(input as FindOptions),
+  ),
+  defineCommand({
     name: 'gesture',
     description: 'Run a structured gesture.',
     inputSchema: fieldsInputSchema(gestureFields),
@@ -164,6 +280,25 @@ export const interactionSemanticCommands = [
     },
   }),
 ] as const;
+
+function defineFieldCommand<
+  const TName extends string,
+  const TFields extends CommandFieldMap,
+  TResult,
+>(
+  name: TName,
+  description: string,
+  fields: TFields,
+  run: (client: AgentDeviceClient, input: InferCommandInput<TFields>) => Promise<TResult>,
+) {
+  return defineCommand({
+    name,
+    description,
+    inputSchema: fieldsInputSchema(fields),
+    readInput: (input) => readFieldInput(input, fields),
+    run,
+  });
+}
 
 function readGestureInput(input: unknown): GestureInput {
   const record = readInputRecord(input);
@@ -246,6 +381,24 @@ function toFillOptions(input: FillInput): FillOptions {
     ...toSelectorSnapshotOptions(input),
     text: input.text,
     delayMs: input.delayMs,
+  };
+}
+
+function toLongPressOptions(input: LongPressInput): LongPressOptions {
+  return {
+    ...commonToClientOptions(input),
+    ...toClientInteractionTarget(input.target),
+    ...toSelectorSnapshotOptions(input),
+    durationMs: input.durationMs,
+  };
+}
+
+function toGetOptions(input: GetInput): GetOptions {
+  return {
+    ...commonToClientOptions(input),
+    ...toClientElementTarget(input.target),
+    ...toSelectorSnapshotOptions(input),
+    format: input.format,
   };
 }
