+++ src/daemon/selector-runtime.ts
@@ -4,7 +4,8 @@ import type {
   BackendSnapshotResult,
 } from '../backend.ts';
 import { createAgentDevice } from '../runtime.ts';
-import { parseWaitPositionals, type WaitParsed } from '../commands/semantic-grammar.ts';
+import { parseWaitPositionals } from '../commands/cli-grammar/capture.ts';
+import type { WaitParsed } from '../commands/cli-grammar/types.ts';
 import { isCommandSupportedOnDevice } from '../core/capabilities.ts';
 import { resolveTargetDevice, type CommandFlags } from '../core/dispatch.ts';
 import { isApplePlatform } from '../utils/device.ts';
