--- src/mcp/catalog.ts
@@ -1,7 +1,0 @@
-import { listSemanticCommandTools } from './semantic-tools.ts';
-
-export const MCP_SERVER_NAME = 'agent-device';
-
-export function listTools(): unknown[] {
-  return listSemanticCommandTools();
-}
