+++ src/utils/cli-command-schema-types.ts
@@ -1,0 +1,14 @@
+import type { CliFlags, FlagKey } from './cli-flags.ts';
+
+export type CommandSchema = {
+  helpDescription: string;
+  summary?: string;
+  positionalArgs?: readonly string[];
+  allowsExtraPositionals?: boolean;
+  allowedFlags?: readonly FlagKey[];
+  defaults?: Partial<CliFlags>;
+  usageOverride?: string;
+  listUsageOverride?: string;
+};
+
+export type CommandSchemaOverride = Partial<CommandSchema>;
